moved macros to sips/macros/
