+++ sips/lines/bov/utils/bov_utils.py
@@ -7,9 +7,11 @@ import requests as r
 
 # from pydash import at
 
-import sips.h.macros as m
+
 import sips.h.openers as o
-from sips.lines.bov import new_lines as nl
+
+from sips.macros import macros as m
+from sips.lines import lines as ll
 from sips.lines.bov.utils import bov_utils as u
 
 HEADERS = {'User-Agent': 'Mozilla/5.0'}
@@ -478,4 +480,4 @@ def bov_all_dict():
 
 if __name__ == "__main__":
     
-    nl.main()
+    ll.main()
