+++ sips/lines/espn/espn_api.py
@@ -1,6 +1,7 @@
 import requests as r
 
-from sips.h import macros as m
+from sips.macros import macros as m
+
 from sips.lines import collate
 
 espn = 'http://site.api.espn.com/apis/site/v2/sports/'
